requirements.txt
streamlit
transformers
torch
Pillow
opencv-python-headless
timm